+++ SpotifyRecorder/Properties/AssemblyInfo.cs
@@ -8,9 +8,9 @@ using System.Windows;
 // Attribute gesteuert. Ändern Sie diese Attributwerte, um die Informationen zu ändern,
 // die einer Assembly zugeordnet sind.
 [assembly: AssemblyTitle("SpotifyRecorder")]
-[assembly: AssemblyDescription("")]
+[assembly: AssemblyDescription("This application can be used to record songs from Spotify.")]
 [assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("")]
+[assembly: AssemblyCompany("Markus Scheich")]
 [assembly: AssemblyProduct("SpotifyRecorder")]
 [assembly: AssemblyCopyright("Copyright ©  2019")]
 [assembly: AssemblyTrademark("")]